Allowed `nowarning=true skills to be autoaccepted and fixed Rush's targeting.
Rush doesn’t need nowarning as it’s impossible to rush yourself because you don’t meet the minimum distance.
Rush doesn’t need nolock as it's no longer free-targeted but can only target creatures.

If the automatically calculated target *is* the player, the ability is almost certain to be a heal, buff, or other thing that makes sense to apply to an ally, and the autotargeting should proceed.
If the automatically calculated target *is not* the player, no self-target warning would trigger anyway, and the autotargeting should proceed.