Allow cancelling Temporal Form's anomaly talents
Added a doAction function to all anomalies which gets wrapped by the
action function and adds a parameter to specify whether targeting
can be cancelled. The default is for this to be true, if the talent
gets used directly by the player, like when in Temporal Form.

For the special cases related to Twist Fate, cancelling is disabled like
it was in the past, even while in Temporal Form.